requirements.txt
pillow==10.2.0
pypinyin==0.51.0
